Move PARSE_FROM_SOURCE scripts into FieldType classes
