+++ server/src/main/java/org/elasticsearch/script/DateFieldScript.java
@@ -13,7 +13,6 @@ import org.elasticsearch.common.time.DateFormatter;
 import org.elasticsearch.search.lookup.SearchLookup;
 
 import java.util.Map;
-import java.util.Objects;
 
 public abstract class DateFieldScript extends AbstractLongFieldScript {
     public static final ScriptContext<Factory> CONTEXT = newContext("date", Factory.class);
@@ -28,25 +27,6 @@ public abstract class DateFieldScript extends AbstractLongFieldScript {
     public interface LeafFactory {
         DateFieldScript newInstance(LeafReaderContext ctx);
     }
-
-    static final Factory PARSE_FROM_SOURCE = (field, params, lookup, formatter) -> (LeafFactory) ctx -> new DateFieldScript(
-        field,
-        params,
-        lookup,
-        formatter,
-        ctx
-    ) {
-        @Override
-        public void execute() {
-            for (Object v : extractFromSource(field)) {
-                try {
-                    emit(formatter.parseMillis(Objects.toString(v)));
-                } catch (Exception e) {
-                    // ignore
-                }
-            }
-        }
-    };
 
     private final DateFormatter formatter;
