+++ server/src/main/java/org/elasticsearch/script/GeoPointFieldScript.java
@@ -10,12 +10,8 @@ package org.elasticsearch.script;
 
 import org.apache.lucene.document.LatLonDocValuesField;
 import org.apache.lucene.index.LeafReaderContext;
-import org.elasticsearch.common.geo.GeoPoint;
-import org.elasticsearch.common.geo.GeoUtils;
-import org.elasticsearch.common.xcontent.support.XContentMapValues;
 import org.elasticsearch.search.lookup.SearchLookup;
 
-import java.util.List;
 import java.util.Map;
 
 import static org.apache.lucene.geo.GeoEncodingUtils.encodeLatitude;
@@ -38,43 +34,6 @@ public abstract class GeoPointFieldScript extends AbstractLongFieldScript {
     public interface LeafFactory {
         GeoPointFieldScript newInstance(LeafReaderContext ctx);
     }
-
-    static final Factory PARSE_FROM_SOURCE = (field, params, lookup) -> (LeafFactory) ctx -> new GeoPointFieldScript(
-        field,
-        params,
-        lookup,
-        ctx
-    ) {
-        private final GeoPoint scratch = new GeoPoint();
-
-        @Override
-        public void execute() {
-            try {
-                Object value = XContentMapValues.extractValue(field, leafSearchLookup.source().source());
-                if (value instanceof List<?>) {
-                    List<?> values = (List<?>) value;
-                    if (values.size() > 0 && values.get(0) instanceof Number) {
-                        parsePoint(value);
-                    } else {
-                        for (Object point : values) {
-                            parsePoint(point);
-                        }
-                    }
-                } else {
-                    parsePoint(value);
-                }
-            } catch (Exception e) {
-                // ignore
-            }
-        }
-
-        private void parsePoint(Object point) {
-            if (point != null) {
-                GeoUtils.parseGeoPoint(point, scratch, true);
-                emit(scratch.lat(), scratch.lon());
-            }
-        }
-    };
 
     public GeoPointFieldScript(String fieldName, Map<String, Object> params, SearchLookup searchLookup, LeafReaderContext ctx) {
         super(fieldName, params, searchLookup, ctx);
