+++ server/src/main/java/org/elasticsearch/script/StringFieldScript.java
@@ -35,22 +35,6 @@ public abstract class StringFieldScript extends AbstractFieldScript {
         StringFieldScript newInstance(LeafReaderContext ctx);
     }
 
-    static final Factory PARSE_FROM_SOURCE = (field, params, lookup) -> (LeafFactory) ctx -> new StringFieldScript(
-        field,
-        params,
-        lookup,
-        ctx
-    ) {
-        @Override
-        public void execute() {
-            for (Object v : extractFromSource(field)) {
-                if (v != null) {
-                    emit(v.toString());
-                }
-            }
-        }
-    };
-
     private final List<String> results = new ArrayList<>();
     private long chars;
